Cambiando la opcion de requerir obligatoriamente el token JWT para solicitar los libros a opcional con la desventaja de no poder usar los filtros @

+++ backend/src/books/books.py
@@ -2,8 +2,10 @@ from flask import Blueprint, jsonify, request
 from app import mongo
 from flask_pymongo import ObjectId
 from schemas.books import validate_book
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from utils.decorators import user_access_required
+from utils.users import get_user_id
+
 
 books_blueprint = Blueprint('books', __name__)
 
@@ -47,20 +49,43 @@ def createBook(user_id):
     
 
 @books_blueprint.route('/books', methods = ['GET'])
-@jwt_required()
-@user_access_required(action='get',name='retrieved',pass_user_id = True)
-def getBooks(user_id):
+@jwt_required(optional=True)
+def getBooks():
+    
+    user_id = get_user_id(get_jwt_identity())
+    aditional = None
+    
     poster = request.args.get('poster')
     if not poster:
         books = mongo.db.books.find({})
     else:
-        if poster == '@me':
-            books = mongo.db.books.find({'user_id': user_id})
-        elif poster == '@others':
-            books = mongo.db.books.find({'user_id': {'$nin': [user_id]}})   
-        else:
+        if not poster.startswith('@'):
             user_id = mongo.db.users.find_one({'username': poster}).get('_id')
             books = mongo.db.books.find({'user_id': user_id})
+        else:
+            if not user_id:
+                return jsonify({'msg': 'poster @ filters are not allowed if JWT token is not provided',
+                        'status': {
+                            'name': 'bad_request',
+                            'action': 'get',
+                            'get': False
+                        }
+                        }), 400
+            if poster == '@me':
+                books = mongo.db.books.find({'user_id': user_id})
+            elif poster == '@others':
+                books = mongo.db.books.find({'user_id': {'$nin': [user_id]}})
+            elif poster == '@all':
+                books = mongo.db.books.find({})  
+            else:
+                return jsonify({'msg': f'unknow @ filter "{poster}"',
+                                'status':{
+                                    'name': 'bad_request',
+                                    'action': 'get',
+                                    'get': False
+                                }
+                                }), 400
+            
     return jsonify({
         'msg': 'Books retrieved',
         'status': {
@@ -74,11 +99,10 @@ def getBooks(user_id):
             'description': book.get('description'),
             'author': book.get('author')
             },books))
-    })
+        })
 
 @books_blueprint.route('/books/<id>', methods = ['GET'])
-@jwt_required()
-@user_access_required(action='get',name='retrieved',pass_user_id = True)
+@jwt_required(optional=True)
 def getBook(id):
     book = mongo.db.books.find_one({'_id': ObjectId(id)})
     if not book:
@@ -107,7 +131,7 @@ def getBook(id):
 @jwt_required()
 @user_access_required(action='delete',name='not_found',pass_user_id = True)
 def deleteBook(id,user_id):
-    book = mongo.db.books.find_one({'user_id': user_id,'_id': ObjectId(id)})
+    book = mongo.db.books.find_one({'_id': ObjectId(id)})
     if not book:
         return jsonify({
             'msg': 'Book not found',
@@ -117,6 +141,16 @@ def deleteBook(id,user_id):
                 'delete': False
                 }   
             })
+    if book.get('user_id') != user_id:
+        return jsonify({
+            'msg': 'This book does not belong to you',
+            'status': {
+                'name': 'not_authorized',
+                'action': 'delete',
+                'delete': False
+                }   
+            })
+    
     
     mongo.db.books.delete_one({'user_id': user_id,'_id': ObjectId(id)})
     return jsonify({'msg': 'Book deleted',
@@ -138,7 +172,7 @@ def updateBook(id,user_id):
             'update': False
         }})
     
-    book = mongo.db.books.find_one({'user_id': user_id,'_id': ObjectId(id)})
+    book = mongo.db.books.find_one({'_id': ObjectId(id)})
     if not book:
         return jsonify({
             'msg': 'Book not found',
@@ -148,6 +182,15 @@ def updateBook(id,user_id):
                 'delete': False
                 }   
             })
+    if book.get('user_id') != user_id:
+        return jsonify({
+            'msg': 'This book does not belong to you',
+            'status': {
+                'name': 'not_authorized',
+                'action': 'update',
+                'update': False
+            }
+        })
     
     if not validate_book(book):
         return jsonify({'msg': 'Invalid book','status': {
